% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-form.R
\name{dt.changeForm}
\alias{dt.changeForm}
\title{Make changes to the parameters specified in initial formula list}
\usage{
dt.changeForm(
  model,
  var,
  type = c("add", "remove"),
  category = c("exposure", "outcome", "covariates", "cohorts")
)
}
\arguments{
\item{model}{list containing four elements (exposure, outcome, covariates, cohort)}

\item{var}{variable(s) to add or remove}

\item{type}{whether to add or remove variables}

\item{category}{list element to modify. Must be one of the four elements specified above.}
}
\value{
runs regression model
}
\description{
Make changes to the parameters specified in initial formula list
}
