% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-table.R
\name{dt.cohortTable}
\alias{dt.cohortTable}
\title{Make a summary table with cohort information}
\usage{
dt.cohortTable(ref, dataset, conns)
}
\arguments{
\item{ref}{path to .csv file containing information on cohorts. At a
minimum must include a column 'cohort' with cohorts in the same format as
the connections object.}

\item{dataset}{Character giving the name of the serverside dataset from which
to calculate cohort ns}

\item{conns}{DataSHIELD connections object.}
}
\value{
Tibble containing same columns as input .csv plus column 'cohort_ns'.
}
\description{
Make a summary table with cohort information
}
