% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-complete-cases.R
\name{dt.defineCompleteCase}
\alias{dt.defineCompleteCase}
\title{Defines whether or not each subject is a complete case based on provided variables}
\usage{
dt.defineCompleteCase(df = NULL, vars = NULL, newobj = NULL, conns = NULL)
}
\arguments{
\item{df}{serverside dataframe to use}

\item{vars}{vector of vars to define completeness}

\item{newobj}{name of new object}

\item{conns}{datashield connections}
}
\value{
creates binary serverside variable indicating missingness
}
\description{
Defines whether or not each subject is a complete case based on provided variables
}
