% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest-data-slma.R
\name{dt.forestDataSLMA}
\alias{dt.forestDataSLMA}
\title{Function to extract model coefficients into tibble which can be used
to make forest plot}
\usage{
dt.forestDataSLMA(obj, mod)
}
\arguments{
\item{obj}{object returned by ds.glmSLMA}

\item{mod}{list model in standard format. This is just use to get the cohorts
which were included in the model. Hopefully in future these will be
extractable from the slma object and this function can be simplified}
}
\value{
tibble in long format with coefficients by study and combined
}
\description{
Function to extract model coefficients into tibble which can be used
to make forest plot
}
