% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-glm-form.R
\name{dt.makeGlmForm}
\alias{dt.makeGlmForm}
\title{Converts a list of formula components into a formula which can be used in
a glm.}
\usage{
dt.makeGlmForm(
  x,
  type = c("ipd", "slma"),
  dummy_suff = "_dummy",
  data = "analysis_df"
)
}
\arguments{
\item{x}{list containing 4 elements: exposure, outcome, covariates, cohorts}

\item{type}{type of model formula will be used for: either "ipd" or "slma"}

\item{dummy_suff}{the suffix used by dummy variables}

\item{data}{the dataframe to be used in the models}
}
\value{
a string of the created formula
}
\description{
Converts a list of formula components into a formula which can be used in
a glm.
}
