% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-to-miss.R
\name{dt.modToMiss}
\alias{dt.modToMiss}
\title{Takes a list of formula components and extracts those needed for missingness diagnostics}
\usage{
dt.modToMiss(x)
}
\arguments{
\item{x}{list containing 4 elements: exposure, outcome, covariates, cohorts}
}
\value{
a string of the created formula
}
\description{
Takes a list of formula components and extracts those needed for missingness diagnostics
}
