% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-wrap.R
\name{dt.glmWrap}
\alias{dt.glmWrap}
\title{Wrapper for the glm function to save repeating arguments when using many times in the same script}
\usage{
dt.glmWrap(mod, type = c("ipd", "slma"), df = "analysis_df", conns = NULL)
}
\arguments{
\item{mod}{output from function ds.makeGlmForm}

\item{type}{type of analysis (either "ipd" or "slma")}

\item{df}{serverside dataframe}

\item{conns}{datashield connections object}
}
\value{
runs regression model
}
\description{
Wrapper for the glm function to save repeating arguments when using many times in the same script
}
